Move graph in components section
