+++ frontend/app/package.json
@@ -4,11 +4,11 @@
     "private": true,
     "proxy": "https://pib-server.azurewebsites.net",
     "dependencies": {
-        "@chakra-ui/react": "^1.6.5",
-        "@emotion/react": "^11.4.0",
+        "@chakra-ui/react": "^1.6.6",
+        "@emotion/react": "^11.4.1",
         "@emotion/styled": "^11.3.0",
-        "@nivo/core": "^0.72.0",
-        "@nivo/line": "^0.72.0",
+        "@nivo/core": "^0.73.0",
+        "@nivo/line": "^0.73.0",
         "dotenv": "^10.0.0",
         "framer-motion": "^4.1.17",
         "moment": "^2.29.1",
@@ -45,9 +45,9 @@
         ]
     },
     "devDependencies": {
-        "@types/jest": "^26.0.24",
-        "@types/node": "^16.3.1",
-        "@types/react": "^17.0.14",
+        "@types/jest": "^27.0.1",
+        "@types/node": "^16.6.1",
+        "@types/react": "^17.0.18",
         "@types/react-dom": "^17.0.9",
         "@types/react-router-dom": "^5.1.8",
         "eslint-config-prettier": "^8.3.0",
